infra(HW05): stack, unit tests (tasty-junit), optimize asymptotic for lookups using `Data.Map`
